feat: 탐지 결과 처리 파이프라인 및 리포지토리 연동 추가

+++ backend/app/infrastructure/db/models/analysis_session.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, String, Integer, DateTime
+from sqlalchemy import Column, String, Integer, DateTime, BigInteger
 from app.infrastructure.db.base import Base
 
 class AnalysisSessionModel(Base):
     __tablename__ = "analysis_session"
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     session_id = Column(String(50), unique=True, nullable=False)
     source_type = Column(String(20), nullable=False)
     source_name = Column(String(255), nullable=True)
